# Conflicts:
#	backend/src/main/java/fr/eseo/tauri/controller/GradeController.java
#	backend/src/main/java/fr/eseo/tauri/controller/TeamController.java
#	backend/src/main/java/fr/eseo/tauri/repository/StudentRepository.java
#	backend/src/main/java/fr/eseo/tauri/seeder/TeamSeeder.java
#	frontend/src/components/pages/EvaluationView.vue

+++ backend/src/main/java/fr/eseo/tauri/model/GradeType.java
@@ -19,7 +19,7 @@ public class GradeType {
     private String name;
 
     @JsonProperty
-    private Integer factor;
+    private Float factor;
 
     @JsonProperty
     private Boolean forGroup;
